Fixed a boxing issue in DynamicTree reducing allocations by another ~400 allocs/sec (4kb/sec).
The only allocations remaining are related to contacts.  In an extreme case (pyramid test) they result in about 200 allocs/sec.  I plan to fix this soon using a pool for contacts & contact edges, but since it's a publicly exposed type I'm going to have to come up with some sort of reuse policy.

git-svn-id: https://box2dxna.svn.codeplex.com/svn@44345 2327b42d-5241-43d6-9e2a-de5ac946f064